Rename HomeHeader to Header, delete original Header component
