--- app/components/home-header.tsx
@@ -1,125 +1,0 @@
-"use client";
-
-import { Suspense, useCallback } from "react";
-import { useAtomValue } from "jotai";
-import {
-  searchedAddressAtom,
-  isSearchCompleteAtom,
-} from "@/atoms/AddressSearchAtom";
-import { useRouter, usePathname } from "next/navigation";
-import { Headings } from "../data/data";
-
-import {
-  Box,
-  Text,
-  HStack,
-  Image,
-  VisuallyHidden,
-  Link,
-  Flex,
-} from "@chakra-ui/react";
-import Heading from "./heading";
-import ReportAddress from "./report-address";
-import SearchBar from "./search-bar";
-import Share from "./share";
-import ShareSkeleton from "./share-skeleton";
-
-export type HazardData = {
-  liquefaction: { exists: boolean; last_updated: string | null } | null;
-  softStory: { exists: boolean; last_updated: string | null } | null;
-  tsunami: { exists: boolean; last_updated: string | null } | null;
-};
-
-const HomeHeader = () => {
-  const pathname = usePathname();
-  const isHome = pathname === "/";
-  const searchedAddress = useAtomValue(searchedAddressAtom);
-  const isSearchComplete = useAtomValue(isSearchCompleteAtom);
-
-  const headingData = Headings.home;
-  const router = useRouter();
-
-  const handleSearchChange = useCallback(
-    (coords: number[], address: string) => {
-      const newUrl = `?address=${encodeURIComponent(address)}&lat=${coords[1]}&lon=${coords[0]}`;
-      router.push(newUrl, { scroll: false });
-    },
-    [router]
-  );
-
-  return (
-    <Box
-      as="header"
-      bgGradient="blue"
-      py={{ base: "4", "2xl": "5" }}
-      px="8"
-      w="full"
-    >
-      <Flex
-        direction={{
-          base: "column",
-          md: "column",
-          lg: "row-reverse",
-        }}
-        justifyContent={"space-between"}
-        alignItems={{ base: "flex-start", xl: "center" }}
-        gap="1.5"
-        mb={{ base: "2" }}
-      >
-        <HStack align="start" gap="1">
-          <Link
-            as={"a"}
-            color="white"
-            href="/"
-            cursor="button"
-            textDecoration={"none"}
-            onClick={(e) => {
-              e.preventDefault();
-              router.push("/");
-            }}
-          >
-            <HStack align="baseline">
-              <Image
-                src="/images/SFSafeHome-fulllogo.svg"
-                alt="SafeHome logo"
-                role="img" // needed for VoiceOver bug: https://bugs.webkit.org/show_bug.cgi?id=216364
-                height="safeHomeLogoHeight"
-                width="safeHomeLogoWidth"
-              />
-              <VisuallyHidden>SafeHome</VisuallyHidden>
-            </HStack>{" "}
-          </Link>
-          <Text textStyle="textPrerelease" layerStyle="prerelease">
-            Beta
-          </Text>
-        </HStack>
-        {isSearchComplete && isHome ? (
-          <ReportAddress searchedAddress={searchedAddress} />
-        ) : (
-          <Heading headingData={headingData} />
-        )}
-      </Flex>
-
-      {isHome ? (
-        <Flex
-          direction={{ base: "column-reverse", xl: "row" }}
-          justifyContent={"space-between"}
-          alignItems={{ base: "flex-start", xl: "center" }}
-        >
-          <Box width={{ base: "full", xl: "fit" }}>
-            <SearchBar onSearchChange={handleSearchChange} />
-          </Box>
-
-          {/* NOTE: This Suspense boundary is being used around a component that utilizes `useSearchParams()` to prevent entire page from deopting into client-side rendering (CSR) bailout as per https://nextjs.org/docs/messages/missing-suspense-with-csr-bailout */}
-          {isSearchComplete ? (
-            <Suspense fallback={<ShareSkeleton />}>
-              <Share />
-            </Suspense>
-          ) : null}
-        </Flex>
-      ) : null}
-    </Box>
-  );
-};
-
-export default HomeHeader;
